Change Boss (#61)
Hello,
my modification of the code allows the employee to change his boss.
Boss changes through textBox. Only an integer (ID Boss) can be entered in the TextBox.
The default value is the current Boss. In the case of a non-existent Boss, the default value is returned.

Thank you for your review.
Luboš Vojáček

The hierarchical structure is treated.